Object movement & random seconds logic
Create initial version of the game, where 3d object a cube falls from a certain place of the screen.
The fall should be on a random principle, e.g 1-4 seconds.
The fall should have speed that can be changed.
Handle the tap event on the screen or the press of “w” for example.